requirements.txt
accelerate==1.0.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
diffusers==0.30.3
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
huggingface-hub==0.25.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.4.1
nltk==3.9.1
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
psutil==6.0.0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
sympy==1.13.3
textblob==0.18.0.post0
tokenizers==0.20.1
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
